+++ lib/ChangeLog.md
@@ -1,5 +1,6 @@
 Connector Changelog
 ===================
 
+* [2023-03-06] added `DiscoveryCms.getAsset()/getAssets()` method and `useDiscoveryAsset()/useDiscoveryAssets()` hooks.
 
 * [2023-02-22] added `DiscoveryCms.getTaxonomy()` method and `useDiscoveryTaxonomy()` hook.